requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
gspread>=5.12.0
oauth2client>=4.1.3
python-dotenv>=1.0.0
requests>=2.31.0
folium>=0.14.0
streamlit-folium>=0.15.0
python-dateutil>=2.8.2
tenacity>=8.2.3

